refactor(team): split manifest+membership+loops out of launcher.go (C14)
Mechanical file split. Three themed files extracted from launcher.go.
This PR pushes launcher.go below the PLAN.md §6 target of <1000
lines (was 1290, now 865).

New files:

* launcher_manifest.go (133 lines, new) — manifest + onboarding
  helpers. isOnboarded, resetManifestToPack/Blueprint,
  resolveRepoRoot (walks up for go.mod or templates/),
  loadRunningSessionMode (queries running broker for session mode).

* launcher_membership.go (197 lines, new) — office membership
  helpers. brokerMemberProviderKind, agentActiveTask,
  officeMembersSnapshot (canonical roster builder),
  isFocusModeEnabled, officeMemberBySlug, agentConfigFromMember
  (member -> agent.AgentConfig transform), activeSessionMembers,
  PackName, AgentCount, filterEnv.

* launcher_loops.go (156 lines, new) — long-running poll-loop
  goroutine entry points. watchChannelPaneLoop +
  channelPaneStatus + captureDeadChannelPane (channel TUI pane
  watcher), primeVisibleAgents (delegate to paneLifecycle +
  replay-catchup), pollOneRelayEventsLoop +
  fetchAndRecordOneRelayEvents (OneRelay backlog poller).

launcher.go: 1290 -> 865 lines (-425, -33%).

Cumulative since main: 4998 -> 865 = -4133 lines (-82.7%).

PLAN.md §6 "definition of done": launcher.go < 1000 lines. ✅

Stacked on refactor/launcher-c13-prompts-mcp (C13).
